requirements.txt
discord.py==1.7.1
aiohttp==3.7.3
pytz==2021.1
pillow==8.3.1
pytesseract==0.3.8